+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Dependency Upgrade
 * Fix #30: Dependency Upgrade
 
+### Bugs
+* Fix #29: Throw an exception if a config is not found
+
 ### 1.0.7 (2025-10-20)
 * Fix#27
   * reverted changes to v1.0.5 for blocking informers from shutdown